Refactoring + added [JsonIgnore] on private properties for Swagger.
