+++ src/Mantasflowers.Contracts/Firebase/GetTokensResponse.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace Mantasflowers.Services.FirebaseService
+namespace Mantasflowers.Contracts.Firebase
 {
     public sealed class GetTokensResponse
     {
@@ -14,9 +14,11 @@ namespace Mantasflowers.Services.FirebaseService
         /// Firebase's different endpoints return 
         /// differently formatted attributes for same values.
         /// </summary>
+        [JsonIgnore]
         [JsonProperty("id_token")]
         private string id_Token { set => IdToken = value; }
 
+        [JsonIgnore]
         [JsonProperty("refresh_token")]
         private string refresh_token { set => RefreshToken = value; }
     }
