+++ src/Mantasflowers.Services/Mantasflowers.Services.csproj
@@ -7,7 +7,10 @@
   <ItemGroup>
     <PackageReference Include="FirebaseAdmin" Version="2.0.0" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Mantasflowers.Contracts\Mantasflowers.Contracts.csproj" />
   </ItemGroup>
 
 </Project>
